Added methods to create different types of cards in js
The new methods create the cards using the previous format with the changes implemented. I'm not 100% sure the state is being loaded correctly yet but that can be fixed later (very minor). Currently the card methods still need to be changed to create the correct divs for the collapsible list so that's next. Some refactoring will be needed at the end to change var naming convention to standard since I kinda forgot about it, will do it after mostly finished to not disrupt workflow