Added necessary code to build the rest of the collapsible list containing the cards to the js file.
The extra divs needed to create the list have been added to the js file. The custom bootstrap function was a bit of a problem and I'm not 100% sure it's working properly since I can't test it yet, but how it's supposed to work and the tags required to add/fix if it doesn't is documented in the code. Also in this commit are small changes to variable names and added a copy of the dummy html page for ease of reference while implementing js to the actual page (delete before master merge).