Added dropdown/checkbox for table_filter (no functions yet)
This adds the actual checkbox into the page. Function to populate the dropdown with the different tables is implemented in a way but actual filtering isn't there yet. The way it's currently parsing the data will become a scaling problem that requires a refactor of the way we're loading data into /waiter/ as a whole but it'll do for now. Noted as todo in code.